Refactored package organisation and added fletcher console script
