+++ src/fletcher/__init__.py
@@ -1,2 +1,1 @@
-
-
+from .fletcher import get_reference_neighbours, compare_queries_to_reference
